+++ src/gui/widgets/interaction/ThreadScrollGeometry.zig
@@ -1,7 +1,7 @@
 //! Current row geometry used to resolve a delivered reading position.
 control: @import("ThreadItemControl.zig"),
-baseline: u32,
+baseline: f64,
 offset: f32,
 maximum: f32,
 step: f32,
-limit: u32,
+limit: f64,
