+++ src/gui/widgets/interaction/ThreadScrollRequest.zig
@@ -1,5 +1,5 @@
 //! An item position copied from the delivered frame that received a disclosure.
 control: @import("ThreadItemControl.zig"),
 sequence: u64 = 0,
-baseline: u32,
+baseline: f64,
 offset: f32,
